Small fixes + Added state for storing previous questions' ids
