+++ connection_router/connection.py
@@ -1,0 +1,22 @@
+import asyncpg
+
+
+# Database connect function
+async def connect_or_create(user, database) -> asyncpg.Connection:
+    try:
+        conn = await asyncpg.connect(user=user, database=database)
+    except asyncpg.InvalidCatalogNameError:
+        # Database does not exist, create it.
+        sys_conn = await asyncpg.connect(
+            database='template1',
+            user='postgres'
+        )
+        await sys_conn.execute(
+            f'CREATE DATABASE "{database}" OWNER "{user}"',
+        )
+        await sys_conn.close()
+
+        # Connect to the newly created database.
+        conn = await asyncpg.connect(user=user, database=database)
+
+    return conn
